chore: refactor rounds in skills to be in correct test folders
