+++ tests/test_skills/test_price_estimation_abci/test_rounds.py
@@ -20,33 +20,15 @@
 """Test the base.py module of the skill."""
 import logging  # noqa: F401
 from types import MappingProxyType
-from typing import Dict, FrozenSet, Optional, Type, cast
+from typing import Dict, FrozenSet, Optional
 
-from packages.valory.skills.abstract_round_abci.base import AbstractRound
 from packages.valory.skills.abstract_round_abci.base import (
     BasePeriodState as PeriodState,
 )
-from packages.valory.skills.abstract_round_abci.base import (
-    BaseTxPayload,
-    CollectSameUntilThresholdRound,
-    StateDB,
-    VotingRound,
-)
+from packages.valory.skills.abstract_round_abci.base import StateDB
 from packages.valory.skills.oracle_deployment_abci.payloads import (
-    DeployOraclePayload,
     RandomnessPayload,
     SelectKeeperPayload,
-)
-from packages.valory.skills.oracle_deployment_abci.rounds import DeployOracleRound
-from packages.valory.skills.oracle_deployment_abci.rounds import (
-    Event as OracleDeploymentEvent,
-)
-from packages.valory.skills.oracle_deployment_abci.rounds import (
-    PeriodState as OracleDeploymentPeriodState,
-)
-from packages.valory.skills.oracle_deployment_abci.rounds import (
-    SelectKeeperOracleRound,
-    ValidateOracleRound,
 )
 from packages.valory.skills.price_estimation_abci.payloads import (
     EstimatePayload,
@@ -67,52 +49,24 @@ from packages.valory.skills.price_estimation_abci.rounds import TxHashRound
 from packages.valory.skills.registration_abci.rounds import (
     BasePeriodState as RegistrationPeriodState,
 )
-from packages.valory.skills.safe_deployment_abci.payloads import DeploySafePayload
-from packages.valory.skills.safe_deployment_abci.rounds import DeploySafeRound
-from packages.valory.skills.safe_deployment_abci.rounds import (
-    Event as SafeDeploymentEvent,
-)
-from packages.valory.skills.safe_deployment_abci.rounds import (
-    PeriodState as SafeDeploymentPeriodState,
-)
-from packages.valory.skills.safe_deployment_abci.rounds import (
-    SelectKeeperSafeRound,
-    ValidateSafeRound,
-)
 from packages.valory.skills.transaction_settlement_abci.payloads import (
-    FinalizationTxPayload,
     ResetPayload,
-    SignaturePayload,
     ValidatePayload,
-)
-from packages.valory.skills.transaction_settlement_abci.rounds import (
-    CollectSignatureRound,
 )
 from packages.valory.skills.transaction_settlement_abci.rounds import (
     Event as TransactionSettlementEvent,
 )
-from packages.valory.skills.transaction_settlement_abci.rounds import FinalizationRound
 from packages.valory.skills.transaction_settlement_abci.rounds import (
     PeriodState as TransactionSettlementPeriodState,
 )
 from packages.valory.skills.transaction_settlement_abci.rounds import (
     RandomnessTransactionSubmissionRound,
-)
-from packages.valory.skills.transaction_settlement_abci.rounds import (
-    ResetRound as ConsensusReachedRound,
-)
-from packages.valory.skills.transaction_settlement_abci.rounds import (
-    SelectKeeperTransactionSubmissionRoundA,
-    SelectKeeperTransactionSubmissionRoundB,
-    ValidateTransactionRound,
     rotate_list,
 )
 
 from tests.test_skills.test_abstract_round_abci.test_base_rounds import (
     BaseCollectDifferentUntilThresholdRoundTest,
     BaseCollectSameUntilThresholdRoundTest,
-    BaseOnlyKeeperSendsRoundTest,
-    BaseVotingRoundTest,
 )
 
 
@@ -229,21 +183,6 @@ def get_most_voted_tx_hash() -> str:
     return "tx_hash"
 
 
-def get_participant_to_signature(
-    participants: FrozenSet[str],
-) -> Dict[str, SignaturePayload]:
-    """participant_to_signature"""
-    return {
-        participant: SignaturePayload(sender=participant, signature="signature")
-        for participant in participants
-    }
-
-
-def get_final_tx_hash() -> str:
-    """final_tx_hash"""
-    return "tx_hash"
-
-
 class TestRandomnessTransactionSubmissionRound(BaseCollectSameUntilThresholdRoundTest):
     """Test RandomnessTransactionSubmissionRound."""
 
@@ -274,61 +213,6 @@ class TestRandomnessTransactionSubmissionRound(BaseCollectSameUntilThresholdRoun
                 exit_event=TransactionSettlementEvent.DONE,
             )
         )
-
-
-class BaseDeployTestClass(BaseOnlyKeeperSendsRoundTest):
-    """Test DeploySafeRound."""
-
-    round_class: Type[AbstractRound]
-    payload_class: Type[BaseTxPayload]
-    update_keyword: str
-
-    def test_run(
-        self,
-    ) -> None:
-        """Run tests."""
-
-        keeper = sorted(list(self.participants))[0]
-        self.period_state = cast(
-            PeriodState,
-            self.period_state.update(most_voted_keeper_address=keeper),
-        )
-
-        test_round = self.round_class(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,  # type: ignore
-                keeper_payloads=self.payload_class(keeper, get_safe_contract_address()),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    **{self.update_keyword: get_safe_contract_address()}
-                ),
-                state_attr_checks=[lambda state: getattr(state, self.update_keyword)],
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-
-class TestDeploySafeRound(BaseDeployTestClass):
-    """Test DeploySafeRound."""
-
-    round_class = DeploySafeRound
-    payload_class = DeploySafePayload
-    update_keyword = "safe_contract_address"
-    _event_class = SafeDeploymentEvent
-    _period_state_class = SafeDeploymentPeriodState
-
-
-class TestDeployOracleRound(BaseDeployTestClass):
-    """Test DeployOracleRound."""
-
-    round_class = DeployOracleRound
-    payload_class = DeployOraclePayload
-    update_keyword = "oracle_contract_address"
-    _event_class = OracleDeploymentEvent
-    _period_state_class = OracleDeploymentPeriodState
 
 
 class TestCollectObservationRound(BaseCollectDifferentUntilThresholdRoundTest):
@@ -469,300 +353,6 @@ class TestTxHashRound(BaseCollectSameUntilThresholdRoundTest):
         )
 
 
-class TestCollectSignatureRound(BaseCollectDifferentUntilThresholdRoundTest):
-    """Test CollectSignatureRound."""
-
-    _period_state_class = TransactionSettlementPeriodState
-    _event_class = TransactionSettlementEvent
-
-    def test_run(
-        self,
-    ) -> None:
-        """Runs tests."""
-
-        test_round = CollectSignatureRound(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,
-                round_payloads=get_participant_to_signature(self.participants),
-                state_update_fn=lambda _period_state, _: _period_state,
-                state_attr_checks=[],
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-    def test_no_majority_event(self) -> None:
-        """Test the no-majority event."""
-        test_round = CollectSignatureRound(self.period_state, self.consensus_params)
-        self._test_no_majority_event(test_round)
-
-
-class TestFinalizationRound(BaseOnlyKeeperSendsRoundTest):
-    """Test FinalizationRound."""
-
-    _period_state_class = TransactionSettlementPeriodState
-    _event_class = TransactionSettlementEvent
-
-    def test_run_success(
-        self,
-    ) -> None:
-        """Runs tests."""
-
-        keeper = sorted(list(self.participants))[0]
-        self.period_state = cast(
-            PeriodState,
-            self.period_state.update(most_voted_keeper_address=keeper),
-        )
-
-        test_round = FinalizationRound(
-            state=self.period_state,
-            consensus_params=self.consensus_params,
-        )
-
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,
-                keeper_payloads=FinalizationTxPayload(sender=keeper, tx_hash="tx_hash"),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    final_tx_hash=get_final_tx_hash()
-                ),
-                state_attr_checks=[lambda state: state.final_tx_hash],
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-    def test_run_failure(
-        self,
-    ) -> None:
-        """Runs tests."""
-
-        keeper = sorted(list(self.participants))[0]
-        self.period_state = cast(
-            PeriodState,
-            self.period_state.update(most_voted_keeper_address=keeper),
-        )
-
-        test_round = FinalizationRound(
-            state=self.period_state,
-            consensus_params=self.consensus_params,
-        )
-
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,
-                keeper_payloads=FinalizationTxPayload(sender=keeper, tx_hash=None),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    final_tx_hash=get_final_tx_hash()
-                ),
-                state_attr_checks=[],
-                exit_event=self._event_class.FAILED,
-            )
-        )
-
-
-class BaseSelectKeeperRoundTest(BaseCollectSameUntilThresholdRoundTest):
-    """Test SelectKeeperTransactionSubmissionRoundA"""
-
-    test_class: Type[CollectSameUntilThresholdRound]
-    test_payload: Type[SelectKeeperPayload]
-
-    _period_state_class = PeriodState
-
-    def test_run(
-        self,
-    ) -> None:
-        """Run tests."""
-
-        test_round = self.test_class(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,
-                round_payloads=get_participant_to_selection(self.participants),
-                state_update_fn=lambda _period_state, _test_round: _period_state.update(
-                    participant_to_selection=MappingProxyType(
-                        dict(get_participant_to_selection(self.participants))
-                    )
-                ),
-                state_attr_checks=[lambda state: state.participant_to_selection.keys()],
-                most_voted_payload="keeper",
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-
-class TestSelectKeeperTransactionSubmissionRoundA(BaseSelectKeeperRoundTest):
-    """Test SelectKeeperTransactionSubmissionRoundA"""
-
-    test_class = SelectKeeperTransactionSubmissionRoundA
-    test_payload = SelectKeeperPayload
-    _event_class = TransactionSettlementEvent
-
-
-class TestSelectKeeperTransactionSubmissionRoundB(BaseSelectKeeperRoundTest):
-    """Test SelectKeeperTransactionSubmissionRoundB."""
-
-    test_class = SelectKeeperTransactionSubmissionRoundB
-    test_payload = SelectKeeperPayload
-    _event_class = TransactionSettlementEvent
-
-
-class TestSelectKeeperSafeRound(BaseSelectKeeperRoundTest):
-    """Test SelectKeeperTransactionSubmissionRoundB."""
-
-    test_class = SelectKeeperSafeRound
-    test_payload = SelectKeeperPayload
-    _event_class = SafeDeploymentEvent
-
-
-class TestSelectKeeperOracleRound(BaseSelectKeeperRoundTest):
-    """Test SelectKeeperTransactionSubmissionRoundB."""
-
-    test_class = SelectKeeperOracleRound
-    test_payload = SelectKeeperPayload
-    _event_class = OracleDeploymentEvent
-
-
-class TestConsensusReachedRound(BaseCollectSameUntilThresholdRoundTest):
-    """Test ConsensusReachedRound."""
-
-    _period_state_class = TransactionSettlementPeriodState
-    _event_class = TransactionSettlementEvent
-
-    def test_runs(
-        self,
-    ) -> None:
-        """Runs tests."""
-
-        test_round = ConsensusReachedRound(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-        next_period_count = 1
-        self._complete_run(
-            self._test_round(
-                test_round=test_round,
-                round_payloads=get_participant_to_period_count(
-                    self.participants, next_period_count
-                ),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    period_count=next_period_count,
-                    participants=self.participants,
-                ),
-                state_attr_checks=[],  # [lambda state: state.participants],
-                most_voted_payload=next_period_count,
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-
-class BaseValidateRoundTest(BaseVotingRoundTest):
-    """Test BaseValidateRound."""
-
-    test_class: Type[VotingRound]
-    test_payload: Type[ValidatePayload]
-
-    def test_positive_votes(
-        self,
-    ) -> None:
-        """Test ValidateRound."""
-
-        test_round = self.test_class(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_voting_round_positive(
-                test_round=test_round,
-                round_payloads=get_participant_to_votes(self.participants),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    participant_to_votes=MappingProxyType(
-                        dict(get_participant_to_votes(self.participants))
-                    )
-                ),
-                state_attr_checks=[lambda state: state.participant_to_votes.keys()],
-                exit_event=self._event_class.DONE,
-            )
-        )
-
-    def test_negative_votes(
-        self,
-    ) -> None:
-        """Test ValidateRound."""
-
-        test_round = self.test_class(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_voting_round_negative(
-                test_round=test_round,
-                round_payloads=get_participant_to_votes(self.participants, vote=False),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    participant_to_votes=MappingProxyType(
-                        dict(get_participant_to_votes(self.participants, vote=False))
-                    )
-                ),
-                state_attr_checks=[],
-                exit_event=self._event_class.NEGATIVE,
-            )
-        )
-
-    def test_none_votes(
-        self,
-    ) -> None:
-        """Test ValidateRound."""
-
-        test_round = self.test_class(
-            state=self.period_state, consensus_params=self.consensus_params
-        )
-
-        self._complete_run(
-            self._test_voting_round_none(
-                test_round=test_round,
-                round_payloads=get_participant_to_votes(self.participants, vote=None),
-                state_update_fn=lambda _period_state, _: _period_state.update(
-                    participant_to_votes=MappingProxyType(
-                        dict(get_participant_to_votes(self.participants, vote=None))
-                    )
-                ),
-                state_attr_checks=[],
-                exit_event=self._event_class.NONE,
-            )
-        )
-
-
-class TestValidateSafeRound(BaseValidateRoundTest):
-    """Test ValidateSafeRound."""
-
-    test_class = ValidateSafeRound
-    test_payload = ValidatePayload
-    _event_class = SafeDeploymentEvent
-    _period_state_class = SafeDeploymentPeriodState
-
-
-class TestValidateOracleRound(BaseValidateRoundTest):
-    """Test ValidateSafeRound."""
-
-    test_class = ValidateOracleRound
-    test_payload = ValidatePayload
-    _event_class = OracleDeploymentEvent
-    _period_state_class = OracleDeploymentPeriodState
-
-
-class TestValidateTransactionRound(BaseValidateRoundTest):
-    """Test ValidateRound."""
-
-    test_class = ValidateTransactionRound
-    test_payload = ValidatePayload
-    _event_class = TransactionSettlementEvent
-    _period_state_class = TransactionSettlementPeriodState
-
-
 def test_rotate_list_method() -> None:
     """Test `rotate_list` method."""
 
@@ -782,8 +372,6 @@ def test_period_states() -> None:
     oracle_contract_address = get_safe_contract_address()
     participant_to_votes = get_participant_to_votes(participants)
     most_voted_tx_hash = get_most_voted_tx_hash()
-    participant_to_signature = get_participant_to_signature(participants)
-    final_tx_hash = get_final_tx_hash()
     participant_to_observations = get_participant_to_observations(participants)
     estimate = get_estimate()
     most_voted_estimate = get_most_voted_estimate()
@@ -810,44 +398,6 @@ def test_period_states() -> None:
     assert period_state.most_voted_keeper_address == most_voted_keeper_address
     assert period_state.participant_to_votes == participant_to_votes
 
-    period_state_ = SafeDeploymentPeriodState(
-        StateDB(
-            initial_period=0,
-            initial_data=dict(
-                participants=participants,
-                participant_to_randomness=participant_to_randomness,
-                most_voted_randomness=most_voted_randomness,
-                participant_to_selection=participant_to_selection,
-                most_voted_keeper_address=most_voted_keeper_address,
-                safe_contract_address=safe_contract_address,
-            ),
-        )
-    )
-    assert period_state_.keeper_randomness == actual_keeper_randomness
-    assert period_state_.most_voted_randomness == most_voted_randomness
-    assert period_state_.most_voted_keeper_address == most_voted_keeper_address
-    assert period_state_.safe_contract_address == safe_contract_address
-
-    period_state__ = OracleDeploymentPeriodState(
-        StateDB(
-            initial_period=0,
-            initial_data=dict(
-                participants=participants,
-                participant_to_randomness=participant_to_randomness,
-                most_voted_randomness=most_voted_randomness,
-                participant_to_selection=participant_to_selection,
-                most_voted_keeper_address=most_voted_keeper_address,
-                safe_contract_address=safe_contract_address,
-                oracle_contract_address=oracle_contract_address,
-            ),
-        )
-    )
-    assert period_state__.keeper_randomness == actual_keeper_randomness
-    assert period_state__.most_voted_randomness == most_voted_randomness
-    assert period_state__.most_voted_keeper_address == most_voted_keeper_address
-    assert period_state__.safe_contract_address == safe_contract_address
-    assert period_state__.oracle_contract_address == oracle_contract_address
-
     period_state____ = RegistrationPeriodState(
         StateDB(
             initial_period=0,
@@ -863,32 +413,6 @@ def test_period_states() -> None:
     assert period_state____.keeper_randomness == actual_keeper_randomness
     assert period_state____.most_voted_randomness == most_voted_randomness
     assert period_state____.most_voted_keeper_address == most_voted_keeper_address
-
-    period_state_____ = TransactionSettlementPeriodState(
-        StateDB(
-            initial_period=0,
-            initial_data=dict(
-                participants=participants,
-                participant_to_randomness=participant_to_randomness,
-                most_voted_randomness=most_voted_randomness,
-                participant_to_selection=participant_to_selection,
-                most_voted_keeper_address=most_voted_keeper_address,
-                safe_contract_address=safe_contract_address,
-                oracle_contract_address=oracle_contract_address,
-                most_voted_tx_hash=most_voted_tx_hash,
-                participant_to_signature=participant_to_signature,
-                final_tx_hash=final_tx_hash,
-            ),
-        )
-    )
-    assert period_state_____.keeper_randomness == actual_keeper_randomness
-    assert period_state_____.most_voted_randomness == most_voted_randomness
-    assert period_state_____.most_voted_keeper_address == most_voted_keeper_address
-    assert period_state_____.safe_contract_address == safe_contract_address
-    assert period_state_____.oracle_contract_address == oracle_contract_address
-    assert period_state_____.most_voted_tx_hash == most_voted_tx_hash
-    assert period_state_____.participant_to_signature == participant_to_signature
-    assert period_state_____.final_tx_hash == final_tx_hash
 
     period_state______ = PriceEstimationPeriodState(
         StateDB(
