+++ tests/test_skills/test_transaction_settlement_abci/test_rounds.py
@@ -1,0 +1,349 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2021 Valory AG
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""Tests for valory/registration_abci skill's rounds."""
+
+import logging  # noqa: F401
+from typing import Dict, FrozenSet, Optional, cast
+
+from packages.valory.skills.abstract_round_abci.base import (
+    BasePeriodState as PeriodState,
+)
+from packages.valory.skills.abstract_round_abci.base import StateDB
+from packages.valory.skills.oracle_deployment_abci.payloads import (
+    RandomnessPayload,
+    SelectKeeperPayload,
+)
+from packages.valory.skills.transaction_settlement_abci.payloads import (
+    FinalizationTxPayload,
+    ResetPayload,
+    SignaturePayload,
+    ValidatePayload,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import (
+    CollectSignatureRound,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import (
+    Event as TransactionSettlementEvent,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import FinalizationRound
+from packages.valory.skills.transaction_settlement_abci.rounds import (
+    PeriodState as TransactionSettlementPeriodState,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import (
+    ResetRound as ConsensusReachedRound,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import (
+    SelectKeeperTransactionSubmissionRoundA,
+    SelectKeeperTransactionSubmissionRoundB,
+    ValidateTransactionRound,
+)
+
+from tests.test_skills.test_abstract_round_abci.test_base_rounds import (
+    BaseCollectDifferentUntilThresholdRoundTest,
+    BaseCollectSameUntilThresholdRoundTest,
+    BaseOnlyKeeperSendsRoundTest,
+)
+from tests.test_skills.test_oracle_deployment_abci.test_rounds import (
+    BaseSelectKeeperRoundTest,
+    BaseValidateRoundTest,
+)
+
+
+MAX_PARTICIPANTS: int = 4
+RANDOMNESS: str = "d1c29dce46f979f9748210d24bce4eae8be91272f5ca1a6aea2832d3dd676f51"
+
+
+def get_participants() -> FrozenSet[str]:
+    """Participants"""
+    return frozenset([f"agent_{i}" for i in range(MAX_PARTICIPANTS)])
+
+
+def get_participant_to_randomness(
+    participants: FrozenSet[str], round_id: int
+) -> Dict[str, RandomnessPayload]:
+    """participant_to_randomness"""
+    return {
+        participant: RandomnessPayload(
+            sender=participant,
+            round_id=round_id,
+            randomness=RANDOMNESS,
+        )
+        for participant in participants
+    }
+
+
+def get_most_voted_randomness() -> str:
+    """most_voted_randomness"""
+    return RANDOMNESS
+
+
+def get_participant_to_selection(
+    participants: FrozenSet[str],
+) -> Dict[str, SelectKeeperPayload]:
+    """participant_to_selection"""
+    return {
+        participant: SelectKeeperPayload(sender=participant, keeper="keeper")
+        for participant in participants
+    }
+
+
+def get_participant_to_period_count(
+    participants: FrozenSet[str], period_count: int
+) -> Dict[str, ResetPayload]:
+    """participant_to_selection"""
+    return {
+        participant: ResetPayload(sender=participant, period_count=period_count)
+        for participant in participants
+    }
+
+
+def get_most_voted_keeper_address() -> str:
+    """most_voted_keeper_address"""
+    return "keeper"
+
+
+def get_safe_contract_address() -> str:
+    """safe_contract_address"""
+    return "0x6f6ab56aca12"
+
+
+def get_participant_to_votes(
+    participants: FrozenSet[str], vote: Optional[bool] = True
+) -> Dict[str, ValidatePayload]:
+    """participant_to_votes"""
+    return {
+        participant: ValidatePayload(sender=participant, vote=vote)
+        for participant in participants
+    }
+
+
+def get_most_voted_tx_hash() -> str:
+    """most_voted_tx_hash"""
+    return "tx_hash"
+
+
+def get_participant_to_signature(
+    participants: FrozenSet[str],
+) -> Dict[str, SignaturePayload]:
+    """participant_to_signature"""
+    return {
+        participant: SignaturePayload(sender=participant, signature="signature")
+        for participant in participants
+    }
+
+
+def get_final_tx_hash() -> str:
+    """final_tx_hash"""
+    return "tx_hash"
+
+
+class TestSelectKeeperTransactionSubmissionRoundA(BaseSelectKeeperRoundTest):
+    """Test SelectKeeperTransactionSubmissionRoundA"""
+
+    test_class = SelectKeeperTransactionSubmissionRoundA
+    test_payload = SelectKeeperPayload
+    _event_class = TransactionSettlementEvent
+
+
+class TestSelectKeeperTransactionSubmissionRoundB(BaseSelectKeeperRoundTest):
+    """Test SelectKeeperTransactionSubmissionRoundB."""
+
+    test_class = SelectKeeperTransactionSubmissionRoundB
+    test_payload = SelectKeeperPayload
+    _event_class = TransactionSettlementEvent
+
+
+class TestFinalizationRound(BaseOnlyKeeperSendsRoundTest):
+    """Test FinalizationRound."""
+
+    _period_state_class = TransactionSettlementPeriodState
+    _event_class = TransactionSettlementEvent
+
+    def test_run_success(
+        self,
+    ) -> None:
+        """Runs tests."""
+
+        keeper = sorted(list(self.participants))[0]
+        self.period_state = cast(
+            PeriodState,
+            self.period_state.update(most_voted_keeper_address=keeper),
+        )
+
+        test_round = FinalizationRound(
+            state=self.period_state,
+            consensus_params=self.consensus_params,
+        )
+
+        self._complete_run(
+            self._test_round(
+                test_round=test_round,
+                keeper_payloads=FinalizationTxPayload(sender=keeper, tx_hash="tx_hash"),
+                state_update_fn=lambda _period_state, _: _period_state.update(
+                    final_tx_hash=get_final_tx_hash()
+                ),
+                state_attr_checks=[lambda state: state.final_tx_hash],
+                exit_event=self._event_class.DONE,
+            )
+        )
+
+    def test_run_failure(
+        self,
+    ) -> None:
+        """Runs tests."""
+
+        keeper = sorted(list(self.participants))[0]
+        self.period_state = cast(
+            PeriodState,
+            self.period_state.update(most_voted_keeper_address=keeper),
+        )
+
+        test_round = FinalizationRound(
+            state=self.period_state,
+            consensus_params=self.consensus_params,
+        )
+
+        self._complete_run(
+            self._test_round(
+                test_round=test_round,
+                keeper_payloads=FinalizationTxPayload(sender=keeper, tx_hash=None),
+                state_update_fn=lambda _period_state, _: _period_state.update(
+                    final_tx_hash=get_final_tx_hash()
+                ),
+                state_attr_checks=[],
+                exit_event=self._event_class.FAILED,
+            )
+        )
+
+
+class TestCollectSignatureRound(BaseCollectDifferentUntilThresholdRoundTest):
+    """Test CollectSignatureRound."""
+
+    _period_state_class = TransactionSettlementPeriodState
+    _event_class = TransactionSettlementEvent
+
+    def test_run(
+        self,
+    ) -> None:
+        """Runs tests."""
+
+        test_round = CollectSignatureRound(
+            state=self.period_state, consensus_params=self.consensus_params
+        )
+
+        self._complete_run(
+            self._test_round(
+                test_round=test_round,
+                round_payloads=get_participant_to_signature(self.participants),
+                state_update_fn=lambda _period_state, _: _period_state,
+                state_attr_checks=[],
+                exit_event=self._event_class.DONE,
+            )
+        )
+
+    def test_no_majority_event(self) -> None:
+        """Test the no-majority event."""
+        test_round = CollectSignatureRound(self.period_state, self.consensus_params)
+        self._test_no_majority_event(test_round)
+
+
+class TestConsensusReachedRound(BaseCollectSameUntilThresholdRoundTest):
+    """Test ConsensusReachedRound."""
+
+    _period_state_class = TransactionSettlementPeriodState
+    _event_class = TransactionSettlementEvent
+
+    def test_runs(
+        self,
+    ) -> None:
+        """Runs tests."""
+
+        test_round = ConsensusReachedRound(
+            state=self.period_state, consensus_params=self.consensus_params
+        )
+        next_period_count = 1
+        self._complete_run(
+            self._test_round(
+                test_round=test_round,
+                round_payloads=get_participant_to_period_count(
+                    self.participants, next_period_count
+                ),
+                state_update_fn=lambda _period_state, _: _period_state.update(
+                    period_count=next_period_count,
+                    participants=self.participants,
+                ),
+                state_attr_checks=[],  # [lambda state: state.participants],
+                most_voted_payload=next_period_count,
+                exit_event=self._event_class.DONE,
+            )
+        )
+
+
+class TestValidateTransactionRound(BaseValidateRoundTest):
+    """Test ValidateRound."""
+
+    test_class = ValidateTransactionRound
+    test_payload = ValidatePayload
+    _event_class = TransactionSettlementEvent
+    _period_state_class = TransactionSettlementPeriodState
+
+
+def test_period_states() -> None:
+    """Test PeriodState."""
+
+    participants = get_participants()
+    participant_to_randomness = get_participant_to_randomness(participants, 1)
+    most_voted_randomness = get_most_voted_randomness()
+    participant_to_selection = get_participant_to_selection(participants)
+    most_voted_keeper_address = get_most_voted_keeper_address()
+    safe_contract_address = get_safe_contract_address()
+    oracle_contract_address = get_safe_contract_address()
+    most_voted_tx_hash = get_most_voted_tx_hash()
+    participant_to_signature = get_participant_to_signature(participants)
+    final_tx_hash = get_final_tx_hash()
+    actual_keeper_randomness = float(
+        (int(most_voted_randomness, base=16) // 10 ** 0 % 10) / 10
+    )
+
+    period_state_____ = TransactionSettlementPeriodState(
+        StateDB(
+            initial_period=0,
+            initial_data=dict(
+                participants=participants,
+                participant_to_randomness=participant_to_randomness,
+                most_voted_randomness=most_voted_randomness,
+                participant_to_selection=participant_to_selection,
+                most_voted_keeper_address=most_voted_keeper_address,
+                safe_contract_address=safe_contract_address,
+                oracle_contract_address=oracle_contract_address,
+                most_voted_tx_hash=most_voted_tx_hash,
+                participant_to_signature=participant_to_signature,
+                final_tx_hash=final_tx_hash,
+            ),
+        )
+    )
+    assert period_state_____.keeper_randomness == actual_keeper_randomness
+    assert period_state_____.most_voted_randomness == most_voted_randomness
+    assert period_state_____.most_voted_keeper_address == most_voted_keeper_address
+    assert period_state_____.safe_contract_address == safe_contract_address
+    assert period_state_____.oracle_contract_address == oracle_contract_address
+    assert period_state_____.most_voted_tx_hash == most_voted_tx_hash
+    assert period_state_____.participant_to_signature == participant_to_signature
+    assert period_state_____.final_tx_hash == final_tx_hash
